Relocated the `Environment` class from the `app.model.cli` package to the `app.cli` package.
