+++ src/app/cli/cmd/config/show.py
@@ -1,6 +1,6 @@
 import click
 from app.cli.entry import pass_environment, format_option
-from app.model.cli import Environment
+from ... import Environment
 from . import group
 
 meta_help: dict = {
