+++ src/app/cli/cmd/image/build.py
@@ -4,7 +4,7 @@ from docker.errors import APIError, BuildError
 from loguru import logger
 from pathlib import Path
 from app.cli.entry import pass_environment, confirm_option
-from app.model.cli import Environment
+from ... import Environment
 from app.model.images import ImageRepo
 from app.util.config import ConfigBuilder
 from . import group
