+++ src/app/cli/cmd/image/clean.py
@@ -4,7 +4,7 @@ from loguru import logger
 from pathlib import Path
 from subprocess import CompletedProcess
 from app.cli.entry import pass_environment, confirm_option
-from app.model.cli import Environment
+from ... import Environment
 from . import group
